moved sortedSummaries to parent, pass to children
