+++ src/components/Header/SecondarySummary.tsx
@@ -2,26 +2,12 @@
 import React from "react";
 import "./secondary-summary.scss";
 import { useUserContext } from "../../contexts/UserContext";
-import { TestType } from "../../enums";
+import { TestSummary } from "../../firebase/firestoreDocumentInterfaces";
 
-const SecondarySummary = () => {
+const SecondarySummary = ({ sortedSummaries }: { sortedSummaries: (TestSummary|undefined)[]}) => {
 	const {userDocument} = useUserContext();
 
-	//#region test summary constants - might use this instead in the future, keeping for now 
-	const words10 = userDocument?.testSummaries.find(summary => summary.testType == TestType.Words && summary.testLength == 10);
-	const words25 = userDocument?.testSummaries.find(summary => summary.testType == TestType.Words && summary.testLength == 25);
-	const words50 = userDocument?.testSummaries.find(summary => summary.testType == TestType.Words && summary.testLength == 50);
-	const words75 = userDocument?.testSummaries.find(summary => summary.testType == TestType.Words && summary.testLength == 75);
-	const words100 = userDocument?.testSummaries.find(summary => summary.testType == TestType.Words && summary.testLength == 100);
-
-	const timed15 = userDocument?.testSummaries.find(summary => summary.testType == TestType.Time && summary.testLength == 15);
-	const timed30 = userDocument?.testSummaries.find(summary => summary.testType == TestType.Time && summary.testLength == 30);
-	const timed45 = userDocument?.testSummaries.find(summary => summary.testType == TestType.Time && summary.testLength == 45);
-	const timed60 = userDocument?.testSummaries.find(summary => summary.testType == TestType.Time && summary.testLength == 60);
-	const timed120 = userDocument?.testSummaries.find(summary => summary.testType == TestType.Time && summary.testLength == 120);
-	//#endregion
-
-	const sortedSummaries = [timed15, timed30, timed45, timed60, timed120, words10, words25, words50, words75, words100];	
+	//#region test summary constants  
 	const timeWordColumnArray = [15, 30, 45, 60, 120, 10, 25, 50, 75, 100];
 
 	return (
